Renaming all scheduler to executor and moving them into their own files
