+++ stlab/utility.hpp
@@ -1,0 +1,32 @@
+/*
+    Copyright 2015 Adobe
+    Distributed under the Boost Software License, Version 1.0.
+    (See accompanying file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+*/
+
+#ifndef SLABFUTURE_UTILITY_HPP
+#define SLABFUTURE_UTILITY_HPP
+
+namespace stlab
+{
+
+/**************************************************************************************************/
+
+template <typename T>
+future<T> make_ready_future(T&& x) {
+    auto p = package<T(T)>(default_executor(), [](auto&& x) { return x; });
+    p.first(x);
+    return p.second;
+}
+
+inline future<void> make_ready_future() {
+    auto p = package<void()>(default_executor(), [](){});
+    p.first();
+    return p.second;
+}
+
+/**************************************************************************************************/
+
+}
+
+#endif //SLABFUTURE_UTILITY_HPP
