+++ test/channel_process_tests.cpp
@@ -53,7 +53,7 @@ BOOST_FIXTURE_TEST_SUITE(int_channel_process_void_functor, channel_test_fixture_
         _receive[0].set_ready();
         std::vector<future<void>> f(10);
         for (auto i = 0; i < 10; ++i) {
-            f.push_back(async(default_scheduler(), [_send = _send[0], i] {
+            f.push_back(async(default_executor(), [_send = _send[0], i] {
                 _send(i);
             }));
         }
@@ -97,7 +97,7 @@ BOOST_FIXTURE_TEST_SUITE(int_channel_process_void_functor, channel_test_fixture_
         _receive[0].set_ready();
         std::vector<future<void>> f(10);
         for (auto i = 0; i < 10; ++i) {
-            f.push_back(async(default_scheduler(), [_send = _send[0], i] {
+            f.push_back(async(default_executor(), [_send = _send[0], i] {
                 _send(i);
             }));
         }
@@ -144,7 +144,7 @@ BOOST_FIXTURE_TEST_SUITE(int_channel_process_void_functor, channel_test_fixture_
         _receive[0].set_ready();
         std::vector<future<void>> f(10);
         for (auto i = 0; i < 10; ++i) {
-            f.push_back(async(default_scheduler(), [_send = _send[0], i] {
+            f.push_back(async(default_executor(), [_send = _send[0], i] {
                 _send(i);
             }));
         }
@@ -263,7 +263,7 @@ BOOST_AUTO_TEST_CASE(int_channel_process_with_two_steps_timed_wo_timeout) {
     sender<int> send;
     receiver<int> receive;
 
-    std::tie(send, receive) = channel<int>(default_scheduler());
+    std::tie(send, receive) = channel<int>(default_executor());
 
     auto check = receive | timed_sum(2) | [&](int x) { result = x; };
