+++ test/channel_tests.cpp
@@ -210,7 +210,7 @@ BOOST_AUTO_TEST_CASE(int_channel_one_value_different_buffer_sizes) {
     for (auto bs : { 0,1,2,10 }) {
         stlab::sender<int> send;
         stlab::receiver<int> receive;
-        std::tie(send, receive) = stlab::channel<int>(stlab::default_scheduler());
+        std::tie(send, receive) = stlab::channel<int>(stlab::default_executor());
         std::atomic_int result{ 0 };
 
         auto check = receive | stlab::buffer_size(bs) & [&](int x) { result += x; };
@@ -232,7 +232,7 @@ BOOST_AUTO_TEST_CASE(int_channel_two_values_different_buffer_sizes) {
     for (auto bs : { 0,1,2,10 }) {
         stlab::sender<int> send;
         stlab::receiver<int> receive;
-        std::tie(send, receive) = stlab::channel<int>(stlab::default_scheduler());
+        std::tie(send, receive) = stlab::channel<int>(stlab::default_executor());
         std::atomic_int result{ 0 };
 
         auto check = receive | stlab::buffer_size(bs) & [&](int x) { result += x; };
@@ -255,7 +255,7 @@ BOOST_AUTO_TEST_CASE(int_channel_many_values_different_buffer_sizes) {
     for (auto bs : { 0,1,2,10 }) {
         stlab::sender<int> send;
         stlab::receiver<int> receive;
-        std::tie(send, receive) = stlab::channel<int>(stlab::default_scheduler());
+        std::tie(send, receive) = stlab::channel<int>(stlab::default_executor());
         std::atomic_int result{ 0 };
 
         auto check = receive | stlab::buffer_size(bs) & [&](int x) { result += x; };
